Removing print statement from core.py's Star.plot function
There was an annoying print statement that was called everytime the `fleck.Star.plot` function was called, which I assume was initially added for testing purposes but was never removed. I've removed the print statement, but if you want to keep it can you add a `verbose` kind of argument to the function to enable/disable this print statement?